Restructure modules (#69)
* Restructure modules

* Fix missing character in Message import

* Fix docs paths

* More documentation fixes

* More documentation clean-ups

* Fix example in README.md
