+++ p2panda-rs/src/identity/key_pair.rs
@@ -3,23 +3,12 @@ use std::convert::TryFrom;
 use ed25519_dalek::{Keypair as Ed25519Keypair, PublicKey, SecretKey, Signature, Signer};
 use rand::rngs::OsRng;
 use serde::{Deserialize, Serialize};
-use thiserror::Error;
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen::prelude::wasm_bindgen;
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen::JsValue;
 
-/// Custom error types for key-pairs.
-#[derive(Error, Debug)]
-pub enum KeyPairError {
-    /// Handle errors from `ed25519` crate.
-    #[error(transparent)]
-    Ed25519(#[from] ed25519_dalek::ed25519::Error),
-
-    /// Handle errors from `hex` crate.
-    #[error(transparent)]
-    HexEncoding(#[from] hex::FromHexError),
-}
+use crate::identity::KeyPairError;
 
 /// Ed25519 key pair for authors to sign bamboo entries with.
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
@@ -44,7 +33,7 @@ impl KeyPair {
     /// ```
     /// # extern crate p2panda_rs;
     /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use p2panda_rs::key_pair::KeyPair;
+    /// use p2panda_rs::identity::KeyPair;
     ///
     /// // Generate new Ed25519 key pair
     /// let key_pair = KeyPair::new();
@@ -76,12 +65,12 @@ impl KeyPair {
     /// ```
     /// # extern crate p2panda_rs;
     /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use p2panda_rs::key_pair::KeyPair;
+    /// use p2panda_rs::identity::KeyPair;
     ///
     /// // Generate new Ed25519 key pair
     /// let key_pair = KeyPair::new();
     ///
-    /// // Derive a key pair from a private key 
+    /// // Derive a key pair from a private key
     /// let key_pair_derived = KeyPair::from_private_key(key_pair.private_key())?;
     ///
     /// assert_eq!(key_pair.public_key_bytes(), key_pair_derived.public_key_bytes());
@@ -134,7 +123,7 @@ impl KeyPair {
     /// ```
     /// # extern crate p2panda_rs;
     /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// use p2panda_rs::key_pair::KeyPair;
+    /// use p2panda_rs::identity::KeyPair;
     ///
     /// // Generate new Ed25519 key pair
     /// let key_pair = KeyPair::new();
